# application/services/custom_services/products/src/models/entitymodels/products.ts
export interface products 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   qunatity: { type: String },
   price: { type: String },
   offer_price: { type: String },
   description: { type: String },
   prod_images: { type: String },
   imagess: { type: String },
   pros: { type: String }
}