# application/services/custom_services/products/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: sampletesting open api specification
   description: products API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8011
     description: products
tags:
- name: products
  description: crud products 
paths:
   "/products/{id}":
      delete:
         tags:
         -  products
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  products
         description: GpGetEntityById
         parameters:
         -  in: "path"
            description: GpGetEntityById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/products/get/search":
      get:
         tags:
         -  products
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/products/get/update":
      put:
         tags:
         -  products
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/products"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/products":
      put:
         tags:
         -  products
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/products"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  products
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  products
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/products"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/products/userid/created_by":
      get:
         tags:
         -  products
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      products:
         type: object
         properties:
            name:
               type: "string"
               description: Description
            qunatity:
               type: "string"
               description: Description
            price:
               type: "string"
               description: Description
            offer_price:
               type: "string"
               description: Description
            description:
               type: "string"
               description: Description
            prod_images:
               type: "string"
               description: Description
            imagess:
               type: "string"
               description: Description
            pros:
               type: "string"
               description: Description